requirements.txt
fastapi
uvicorn
transformers
torch
langdetect
requests
python-dotenv
wikipedia
